+++ src/javascript/defaults/dialogs/about.js
@@ -1,11 +1,14 @@
 import Dialog from '../../constructors/dialog'
 import packageJSON from '../../../../package.json'
+import buildJSON from '../../../../assets/build.json'
 
 function About(){
 	const DialogInstance = new Dialog({
 		title:'About',
 		content:`
 			Graviton v${packageJSON.version} 
+
+			Build date: ${buildJSON.date}
 
 			Author: Marc Espín Sanz
 		`,